+++ src/connect/components/revision-history/timeline/index.ts
@@ -1,0 +1,1 @@
+export * from './timeline';
