--- src/connect/components/revision-history/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './make-revisions-by-date';
